==> 01.pkl_to_parquet.ipynb <==
import pickle
import pandas as pd 

## 01. 준비된 pkl 파일 불러오기 

data = pd.read_pickle('my_data/all_docs.pkl')

len(data)

data[0]

doc_id_list = []

df = pd.DataFrame(columns=["doc_id", "contents", "metadata"])


for i in range(len(data)):
    meta = data[i].metadata
    contents = data[i].page_content
    
    meta_keys = list(meta.keys())
    filename = meta['filename']
    
    if 'chunk' in meta_keys:
        page = meta['chunk']
    else:
        page = meta['page_no']
    
    filename = ' / '.join(filename.split('/')[1:])
    doc_id = filename + ' - ' + str(page)
    
    df.loc[i] = [doc_id, contents, meta]
    

df.tail(10)

df.iloc[-66:].head()

## 02. parquet 파일로 저장

df.iloc[-66:].to_parquet('my_data/my_corpus.parquet')

# df.to_excel('my_data/my_corpus.xlsx', index=False)

## 03. parquet 파일 잘 저장되었는지 불러보기

path = './my_data/my_corpus.parquet'

# make_qa.py 파일로 만듬 : 질문과 응답 쌍을 만듬 , 어떤 문서가 관련이 있는지 
df = pd.read_parquet(path)

df.head()

df.tail()

len(df)

df['metadata'].iloc[0]

from tqdm import tqdm
from langchain.schema import Document

documents = []

for i in range(len(df)):
    meta = df['metadata'].iloc[i]
    content = df['contents'].iloc[i]
    documents.append(Document(metadata=meta, page_content=content))




from utils.RAGPipeLine import Ragpipeline

pipeline = Ragpipeline()

vectorDB = pipeline.init_vectorDB('./database_new')


vectorDB.add_documents(documents)

retriever = pipeline.init_retriever()

query = 'FSSAI 라이선스 발급 절차는 몇 단계로 이루어져 있나요?'

docs = retriever.get_relevant_documents(query)

docs

docs[0].metadata
