==> 03.00.make_experiment_data_for_RAGAS.ipynb <==
### 만든 실험 데이터셋 불러오기

import pandas as pd 

def get_dataset(path):
    # make_qa.py 파일로 만듬 : 질문과 응답 쌍을 만듬 , 어떤 문서가 관련이 있는지 
    df = pd.read_parquet(path)

    df['retrieval_gt'] = df['retrieval_gt'].apply(lambda x: x[0][0])
    df['generation_gt'] = df['generation_gt'].apply(lambda x: x[0])
    qa_data = df[['query', 'generation_gt']]
    qa_data.columns = ['question', 'ground_truth']
    qa_data = qa_data.loc[~qa_data['ground_truth'].str.contains('해당 자료와 관련이 없습니다.')]
    return qa_data

### 1. QA 실험데이터셋 만들기

df = pd.read_parquet('./my_data/my_qa.parquet')

df


path = './my_data/my_qa.parquet'

df = get_dataset(path)
df

len(df)

df.to_excel('./experiment_dataset/qa_data.xlsx', index=False)

### 2. Title 실험데이터셋 만들기


path = './my_data/my_title.parquet'

df = get_dataset(path)
df.head()

df.to_excel('./experiment_dataset/title_data.xlsx', index=False)

### 3. Post 실험데이터셋 만들기


path = './my_data/my_post.parquet'

df = get_dataset(path)
df.head()

df.to_excel('./experiment_dataset/post_data.xlsx', index=False)
