==> 04.00.qa_experiment_RAGAS.ipynb <==
import pandas as pd 

## 만든 QA 질문-답변정답 데이터셋 불러오기

qa_data = pd.read_excel('./experiment_dataset/qa_data.xlsx')

qa_data.head()

print(len(qa_data))

### 실험을 위한 RAG chain 

from utils.RAGPipeLine import Ragpipeline
pipeline = Ragpipeline()
VectorDB = pipeline.init_vectorDB('./database_new')

# setting0

VectorDB_empty = pipeline.init_vectorDB('./empty')

retriever0 = VectorDB_empty.as_retriever(
            search_kwargs = {"k": 3},
            search_type   = "similarity"
        )

retriever1 = VectorDB.as_retriever(
            search_kwargs = {"k": 3},
            search_type   = "similarity"
        )

retriever2 = VectorDB.as_retriever(
            search_kwargs = {"k": 3},
            search_type   = "mmr"
        )

# basic_retriever = pipeline.init_retriever()
basic_retriever = VectorDB.as_retriever(
            search_kwargs = {"score_threshold": 0.5, "k": 3},
            search_type   = "similarity_score_threshold"
        )

mq_retriever = pipeline.init_multi_query_retriever()
sq_retriever = pipeline.init_self_query_retriever()
pd_retriever  = pipeline.init_parent_document_retriever()
web_retriever = pipeline.init_web_research_retriever()

query = '2020년 1월부터 8월까지 인도가 한국을 대상으로 개시한 반덤핑 조사는 몇 건인가요?'

docs = retriever0.get_relevant_documents(query)

docs

query = '2020년 1월부터 8월까지 인도가 한국을 대상으로 개시한 반덤핑 조사는 몇 건인가요?'

docs = retriever1.get_relevant_documents(query)

docs

from langchain_openai import ChatOpenAI, OpenAIEmbeddings

from utils.config import config
our_llm = ChatOpenAI(
            model       = config['llm_predictor']['model_name'],
            temperature = config['llm_predictor']['temperature'],

        )

from langchain.chains.combine_documents import create_stuff_documents_chain
from langchain.chains import create_retrieval_chain
from utils.prompt import qa_prompt2    

def set_chain(llm, retriever, prompt):
    
    question_answer_chain = create_stuff_documents_chain(llm, prompt)           # 문서 목록을 가져와서 모두 프롬프트로 포맷한 다음 해당 프롬프트를 LLM에 전달합니다.
    
    rag_chat_chain = create_retrieval_chain(retriever, question_answer_chain)  # 사용자 문의를 받아 리트리버로 전달하여 관련 문서를 가져옵니다. 그런 다음 해당 문서(및 원본 입력)는 LLM으로 전달되어 응답을 생성

    return rag_chat_chain


def start_experiment_and_save(qa_data, llm, retriever, prompt):
    rag_chain = set_chain(llm, retriever, prompt)
    
    qa_data['answer'] = qa_data['question'].apply(lambda x: rag_chain.invoke({'input':x})['answer'])
    qa_data['contexts'] = qa_data['question'].apply(lambda x: [d.page_content for d in retriever.get_relevant_documents(x)])
    # 5분 걸림
    return qa_data

# @ ====================

results = start_experiment_and_save(qa_data, our_llm, retriever0, qa_prompt2)

results.head()

results.to_excel('./experiment_dataset/QA/setting0_retriever0.xlsx', index=False)

### 반복

for i in [3,5,7,10]:
    # setting1 
    retriever1 = VectorDB.as_retriever(
                search_kwargs = {"k": i},
                search_type   = "similarity"
            )

    results = start_experiment_and_save(qa_data, our_llm, retriever1, qa_prompt2)
    results.to_excel('./experiment_dataset/QA/setting1_retriever1_{:03d}.xlsx'.format(i), index=False)
    print(i)

### 2.retriever2 실험 - mmr

for i in [3,5,7,10]:
    j = 3
    m=5
    # setting2
    retriever2 = VectorDB.as_retriever(
            search_kwargs = {'fetch_k': i, "k": j, 'lambda_mult': m/10},
            search_type   = "mmr"
        )

    results = start_experiment_and_save(qa_data, our_llm, retriever2, qa_prompt2)
    results.to_excel('./experiment_dataset/QA/setting2_retriever2_i{:03d}_j{:03d}_m{:03d}.xlsx'.format(i,j,m), index=False)
    print(i, j, m)

for i in [3,5,7,10]:
    for j in [1,3,5,7,10]:
        for m in range(2,9,1):  # 쿼리와의 유사성과 선택된 문서 간의 다양성 사이의 균형을 조절, 람다가 1일때, 유사성만 고려, 람다가 0일때, 다양성만 고려
            if j>i:
                continue
            # setting2
            retriever2 = VectorDB.as_retriever(
                    search_kwargs = {'fetch_k': i, "k": j, 'lambda_mult': m/10},
                    search_type   = "mmr"
                )

            results = start_experiment_and_save(qa_data, our_llm, retriever2, qa_prompt2)
            results.to_excel('./experiment_dataset/QA/setting3_retriever2_i{:03d}_j{:03d}_m{:03d}.xlsx'.format(i,j,m), index=False)
            print(i, j, m)

### 4. our basic_retriever 실험

# Only retrieve documents that have a relevance score
# Above a certain threshold

for i in [3,5,7,10]: 
    for j in range(1,8,1): 
        # setting3 
        retriever4 = VectorDB.as_retriever(
                search_kwargs = {"score_threshold": j/10, "k": i},
                search_type   = "similarity_score_threshold"
            )

        results = start_experiment_and_save(qa_data, our_llm, retriever4, qa_prompt2)
        results.to_excel('./experiment_dataset/QA/setting4_our_retriever_{:03d}_{:03d}.xlsx'.format(i,j), index=False)
        print(i, j)

### 5. multi_query_retriever

# multi_query_retriever는 딱히 인자가 없다.
from langchain.retrievers.multi_query import MultiQueryRetriever


best_retriever1 = VectorDB.as_retriever( search_kwargs = {"k": 5}, search_type = "similarity" ) 

best_retriever2 = VectorDB.as_retriever( search_kwargs = {"score_threshold": 0.5, "k": 3}, search_type = "similarity_score_threshold" )

best_retriever3 = VectorDB.as_retriever( search_kwargs = {'fetch_k': 10, 'k': 5, 'lambda_mult': 0.4}, search_type = "mmr")


for i, retriever in enumerate([best_retriever1, best_retriever2, best_retriever3]):

    mq_retriever = MultiQueryRetriever.from_llm(
        retriever=retriever, llm=our_llm
    )

    results = start_experiment_and_save(qa_data, our_llm, mq_retriever, qa_prompt2)
    results.to_excel('./experiment_dataset/QA/setting4_mq_retriever_{:03d}.xlsx'.format(i+1), index=False)
    print(i)


### 6. self_query_retriever

# self_query_retriever는 딱히 인자가 없다.

results = start_experiment_and_save(qa_data, our_llm, sq_retriever, qa_prompt2)
results.to_excel('./experiment_dataset/QA/setting5_sq_retriever.xlsx', index=False)


### 7. parent_document_retriever

를 사용할 수 없다.

# # self_query_retriever는 딱히 인자가 없다.

# results = start_experiment_and_save(qa_data, our_llm, pd_retriever, qa_prompt2)
# results.to_excel('./experiment_dataset/QA/setting7_pd_retriever.xlsx', index=False)


results.tail()


best_retriever1 = VectorDB.as_retriever( search_kwargs = {"k": 5}, search_type   = "similarity" ) 

best_retriever2 = VectorDB.as_retriever( search_kwargs = {"score_threshold": 0.5, "k": 3}, search_type   = "similarity_score_threshold" )

best_retriever3 = VectorDB.as_retriever(
                search_kwargs = {'fetch_k': 10, 'k': 5, 'lambda_mult': 0.4}, search_type   = "mmr")




for i, retriever in enumerate([best_retriever1, best_retriever2, best_retriever3]):

    results = start_experiment_and_save(qa_data, our_llm, retriever, qa_prompt2)
    results.to_excel('./experiment_dataset/QA/ablation_study_best_retrievers_{:03d}.xlsx'.format(i+1), index=False)
    print(i)
