==> 04.01.qa_experiment_RAGAS.ipynb <==
import pandas as pd 

qa_data = pd.read_excel('./my_data/experiment_data.xlsx')

qa_data.head()

print(len(qa_data))

import ast
for col in ['contexts']:
    qa_data[col] = qa_data[col].apply(ast.literal_eval)

qa_data.info()

from datasets import Dataset 

dataset = Dataset.from_pandas(qa_data)

dataset[0]

from ragas.metrics import (
    faithfulness,
    answer_relevancy,
    context_recall,
    context_precision,
)

import nest_asyncio         # nest_asyncio는 파이썬의 비동기 코드 실행을 도와주는 패키지
nest_asyncio.apply()        # Jupyter 노트북과 같이 이벤트 루프가 이미 실행 중인 환경에서 비동기 코드를 여러 번 실행할 수 있도록 설정하는 데 사용
# 이거 안하면 무조건 에러남 

metrics = [
        faithfulness,
        answer_relevancy,
        context_recall,
        context_precision,
    ]

from ragas import evaluate

result = evaluate(
    dataset,
    metrics=metrics,
    raise_exceptions=False  # Add this line
)

result

df1 = pd.DataFrame(list(result.items()), columns=['Metric', 'setting1'])

df1
