==> 06.00.check_results.ipynb <==
import pandas as pd 
import ast
import nest_asyncio         # nest_asyncio는 파이썬의 비동기 코드 실행을 도와주는 패키지
nest_asyncio.apply()        # Jupyter 노트북과 같이 이벤트 루프가 이미 실행 중인 환경에서 비동기 코드를 여러 번 실행할 수 있도록 설정하는 데 사용
# 이거 안하면 무조건 에러남 
import os 

base_dir = 'experiment_dataset/QA_results'



path = os.path.join(base_dir, 'setting1_retriever1.xlsx')

# make_corpus.py 파일로 만듬 : SimpleDirectoryReader와 TokenTextSplitter로 자른 내용을 넣음 
df = pd.read_excel(path)

max_mean_index = df.loc[df['mean'].idxmax(), 'index']

print(max_mean_index)
print(df['mean'].max())


df



path = os.path.join(base_dir, 'setting2_retriever2.xlsx')

# make_corpus.py 파일로 만듬 : SimpleDirectoryReader와 TokenTextSplitter로 자른 내용을 넣음 
df = pd.read_excel(path)

max_mean_index = df.loc[df['mean'].idxmax(), 'index']

print(max_mean_index)
print(df['mean'].max())


df



path = os.path.join(base_dir, 'setting3_retriever3.xlsx')

# make_corpus.py 파일로 만듬 : SimpleDirectoryReader와 TokenTextSplitter로 자른 내용을 넣음 
df = pd.read_excel(path)

max_mean_index = df.loc[df['mean'].idxmax(), 'index']

print(max_mean_index)
print(df['mean'].max())


df



path = os.path.join(base_dir, 'setting4_our_retriever.xlsx')

# make_corpus.py 파일로 만듬 : SimpleDirectoryReader와 TokenTextSplitter로 자른 내용을 넣음 
df = pd.read_excel(path)

max_mean_index = df.loc[df['mean'].idxmax(), 'index']

print(max_mean_index)
print(df['mean'].max())


df



path = os.path.join(base_dir, 'setting5_mq_retriever.xlsx')

# make_corpus.py 파일로 만듬 : SimpleDirectoryReader와 TokenTextSplitter로 자른 내용을 넣음 
df = pd.read_excel(path)

max_mean_index = df.loc[df['mean'].idxmax(), 'index']

print(max_mean_index)
print(df['mean'].max())


df



path = os.path.join(base_dir, 'setting6_sq_retriever.xlsx')

# make_corpus.py 파일로 만듬 : SimpleDirectoryReader와 TokenTextSplitter로 자른 내용을 넣음 
df = pd.read_excel(path)

max_mean_index = df.loc[df['mean'].idxmax(), 'index']

print(max_mean_index)
print(df['mean'].max())


df
